refactoring uri utils
