+++ src/services/dataSourceEditorProvider.ts
@@ -45,9 +45,9 @@ import { MetaObjectPayload } from "../models/meta";
 import { UDA } from "../models/uda";
 import { getBasename, offerConnectAction } from "../utils/core";
 import { getNonce } from "../utils/getNonce";
-import { getUri } from "../utils/getUri";
 import { MessageKind, Runner, notify } from "../utils/notifications";
 import { parseUDAList } from "../utils/uda";
+import { getUri } from "../utils/uriUtils";
 
 const logger = "dataSourceEditorProvider";
