+++ src/utils/uriUtils.ts
@@ -11,12 +11,21 @@
  * specific language governing permissions and limitations under the License.
  */
 
-import { Uri, Webview } from "vscode";
+import { env, Uri, Webview } from "vscode";
 
 export function getUri(
   webview: Webview,
   extensionUri: Uri,
-  pathList: string[]
+  pathList: string[],
 ) {
   return webview.asWebviewUri(Uri.joinPath(extensionUri, ...pathList));
+}
+
+export async function openUrl(url: string): Promise<void> {
+  const uri: Uri | undefined = Uri.parse(url);
+  if (uri === undefined) {
+    throw Error("Invalid url");
+  } else {
+    env.openExternal(uri);
+  }
 }
